sasa/requirements.txt
google-generativeai
speechrecognition
gtts
pygame
pyaudio
gpiozero
setuptools

